adding final touches

+++ src/main/dnd/business/units/Range.java
@@ -4,7 +4,7 @@ import dnd.business.board.Position;
 
 public class Range {
     public static int range(Position player, Position enemy){
-        return Math.sqrt((player.getX() - enemy.getX())*(player.getX() - enemy.getX()) +
+        return (int) Math.sqrt((player.getX() - enemy.getX())*(player.getX() - enemy.getX()) +
                 (player.getY()-enemy.getY())*(player.getY()-enemy.getY()));
     }
